Fix scaling, refactor initialization and style of qpalm.c
To avoid having to check which function is called, when the data is
loaded etc., the scaling is now implemented at the start of the solve
function. Correspondingly, the data is unscaled at the end, so that in
case of resolving, scaling will have the same effect as before.

Other procedures related to the initialization have been encapsulated in
a new (static) function.

I plan to migrate the coding style (indentation and brackets) to that of
LADEL, which was clearer. In this commit, I already fixed the style in
qpalm.c